+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.TalkbackTutorial">
         <activity
-            android:name=".activities.modules.ScrollingModuleActivity"
+            android:name=".activities.modules.scrolling.ScrollingModuleActivity"
             android:exported="false" />
         <activity
             android:name=".activities.LessonActivity"
@@ -34,28 +34,28 @@
             android:name=".activities.modules.TodoModuleXModuleActivity"
             android:exported="false" />
         <activity
-            android:name=".activities.lesson2.Lesson2Module1P2Fragment"
+            android:name=".activities.modules.explorebytouch.Lesson2Module1P2Fragment"
             android:exported="false" />
         <activity
-            android:name=".activities.lesson2.Lesson2Module1P2Fragment"
+            android:name=".activities.modules.explorebytouch.Lesson2Module1P2Fragment"
             android:exported="false" />
         <activity
-            android:name=".activities.modules.ExploreMenuByTouchActivity"
+            android:name=".activities.modules.explorebytouch.ExploreMenuByTouchActivity"
             android:exported="false" />
         <activity
             android:name=".lessons.modules.ExploreMenuByTouchModule"
             android:exported="false" />
         <activity
-            android:name=".activities.lesson2.Lesson2Module1P2Fragment"
+            android:name=".activities.modules.explorebytouch.Lesson2Module1P2Fragment"
             android:exported="false" />
         <activity
-            android:name=".activities.lesson2.selectsettingmodule.SelectSettingPart1Fragment"
+            android:name=".activities.modules.adjustingmenusliders.SelectSettingPart1Fragment"
             android:exported="false" />
         <activity
-            android:name=".activities.modules.SelectSettingModuleActivity"
+            android:name=".activities.modules.adjustingmenusliders.SelectSettingModuleActivity"
             android:exported="false" />
         <activity
-            android:name=".lessons.modules.lesson2.SelectSettingModule"
+            android:name=".lessons.modules.SelectSettingModule"
             android:exported="false" />
     </application>
