+++ app/src/main/java/com/github/talkbacktutorial/activities/modules/scrolling/ScrollingModuleActivity.kt
@@ -1,31 +1,26 @@
-package com.github.talkbacktutorial.activities.modules
+package com.github.talkbacktutorial.activities.modules.scrolling
 
 import android.content.Intent
-import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.commit
 import com.github.talkbacktutorial.R
 import com.github.talkbacktutorial.activities.MainActivity
-import com.github.talkbacktutorial.activities.lesson2.Lesson2Module1P2Fragment
-import com.github.talkbacktutorial.databinding.ActivityExploreMenuByTouchBinding
+import com.github.talkbacktutorial.databinding.ActivityScrollingModuleBinding
 
-/**
- * Explore menu by touch activity
- * @author Jason Wu
- */
-class ExploreMenuByTouchActivity : AppCompatActivity() {
+class ScrollingModuleActivity : AppCompatActivity() {
 
-    lateinit var binding: ActivityExploreMenuByTouchBinding
+    lateinit var binding: ActivityScrollingModuleBinding
         private set
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         supportActionBar?.hide()
-        this.binding = DataBindingUtil.setContentView(this, R.layout.activity_explore_menu_by_touch)
+        this.binding = DataBindingUtil.setContentView(this, R.layout.activity_scrolling_module)
         supportFragmentManager.commit {
-            replace(R.id.frame, Lesson2Module1P2Fragment.newInstance())
-            addToBackStack("lesson2module1p1")
+            replace(R.id.frame, ScrollingModulePart1Fragment.newInstance())
+            addToBackStack("scrollingModulePart1")
         }
     }
 
@@ -38,4 +33,5 @@ class ExploreMenuByTouchActivity : AppCompatActivity() {
             startActivity(intent)
         }
     }
+
 }
