+++ app/src/main/java/com/github/talkbacktutorial/lessons/modules/ExploreMenuByTouchModule.kt
@@ -2,16 +2,18 @@ package com.github.talkbacktutorial.lessons.modules
 
 import android.content.Context
 import android.content.Intent
-import com.github.talkbacktutorial.activities.modules.ExploreMenuByTouchActivity
+import com.github.talkbacktutorial.activities.modules.explorebytouch.ExploreMenuByTouchActivity
 
 /**
  * Explore menu by touch module
  * @author Jason Wu
  */
 class ExploreMenuByTouchModule : Module() {
-    override val title: String = "Explore Menu By Touch" // E.g. "Pause and Play Media"
+
+    override val title: String = "Explore Menu by Touch" // E.g. "Pause and Play Media"
 
     override fun startActivity(context: Context) {
         context.startActivity(Intent(context, ExploreMenuByTouchActivity::class.java))
     }
+
 }
