+++ app/src/main/java/com/github/talkbacktutorial/lessons/modules/ScrollingModule.kt
@@ -1,11 +1,13 @@
-package com.github.talkbacktutorial.lessons.modules.lesson2
+package com.github.talkbacktutorial.lessons.modules
 
 import android.content.Context
 import android.content.Intent
-import com.github.talkbacktutorial.activities.modules.ScrollingModuleActivity
-import com.github.talkbacktutorial.activities.modules.SelectSettingModuleActivity
-import com.github.talkbacktutorial.lessons.modules.Module
+import com.github.talkbacktutorial.activities.modules.scrolling.ScrollingModuleActivity
 
+/**
+ * A module that teaches the user how to scroll through vertical and horizontal menus.
+ * @author Andre Pham
+ */
 class ScrollingModule : Module() {
 
     override val title: String = "Scrolling Vertical and Horizontal Menus"
